+++ include/interface/export/export_c.h
@@ -46,9 +46,9 @@ class ExportC: public Interface
 		bool addInstanceObject(int instance_id, std::string &&base_object_name) noexcept override;
 		bool addInstanceOfInstance(int instance_id, size_t base_instance_id) noexcept override;
 		bool addInstanceMatrix(int instance_id, Matrix4 &&obj_to_world, float time) noexcept override;
-		int addVertex(Point3 &&vertex, size_t time_step) noexcept override; //!< add vertex to mesh; returns index to be used for addTriangle/addQuad
-		int addVertex(Point3 &&vertex, Point3 &&orco, size_t time_step) noexcept override; //!< add vertex with Orco to mesh; returns index to be used for addTriangle/addQuad
-		void addVertexNormal(Vec3 &&normal, size_t time_step) noexcept override; //!< add vertex normal to mesh; the vertex that will be attached to is the last one inserted by addVertex method
+		int addVertex(Point3 &&vertex, int time_step) noexcept override; //!< add vertex to mesh; returns index to be used for addTriangle/addQuad
+		int addVertex(Point3 &&vertex, Point3 &&orco, int time_step) noexcept override; //!< add vertex with Orco to mesh; returns index to be used for addTriangle/addQuad
+		void addVertexNormal(Vec3 &&normal, int time_step) noexcept override; //!< add vertex normal to mesh; the vertex that will be attached to is the last one inserted by addVertex method
 		bool addFace(std::vector<int> &&vertices, std::vector<int> &&uv_indices) noexcept override;
 		int addUv(Uv &&uv) noexcept override;
 		bool smoothVerticesNormals(std::string &&name, double angle) noexcept override;
